Refactor: create folder to component speaker-section
