+++ pages/speakers/[slug].tsx
@@ -17,7 +17,7 @@
 import { GetStaticProps, GetStaticPaths } from 'next';
 
 import Page from '@components/page';
-import SpeakerSection from '@components/speaker-section';
+import SpeakerSection from '@components/speaker-section/speaker-section';
 import Layout from '@components/layout';
 
 import { getAllSpeakers } from '@lib/cms-api';
